feat(common): made RetryPolicy interface public
made backup_policy public
