+++ google/cloud/bigtable/rpc_backoff_policy.h
@@ -17,7 +17,7 @@
 
 #include "google/cloud/bigtable/internal/rpc_policy_parameters.h"
 #include "google/cloud/bigtable/version.h"
-#include "google/cloud/internal/backoff_policy.h"
+#include "google/cloud/backoff_policy.h"
 #include <grpcpp/grpcpp.h>
 #include <chrono>
 #include <memory>
@@ -99,7 +99,7 @@ class ExponentialBackoffPolicy : public RPCBackoffPolicy {
   std::chrono::milliseconds OnCompletion(grpc::Status const& status) override;
 
  private:
-  using Impl = google::cloud::internal::ExponentialBackoffPolicy;
+  using Impl = google::cloud::ExponentialBackoffPolicy;
   Impl impl_;
 };
