+++ google/cloud/internal/retry_policy.h
@@ -15,6 +15,7 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_RETRY_POLICY_H
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_RETRY_POLICY_H
 
+#include "google/cloud/retry_policy.h"
 #include "google/cloud/status.h"
 #include "google/cloud/version.h"
 #include <chrono>
@@ -26,46 +27,6 @@ inline namespace GOOGLE_CLOUD_CPP_NS {
 namespace internal {
 
 enum class Idempotency { kIdempotent, kNonIdempotent };
-
-/**
- * Define the interface for retry policies.
- */
-class RetryPolicy {
- public:
-  virtual ~RetryPolicy() = default;
-
-  //@{
-  /**
-   * @name Control retry loop duration.
-   *
-   * This functions are typically used in a retry loop, where they control
-   * whether to continue, whether a failure should be retried, and finally
-   * how to format the error message.
-   *
-   * @code
-   * std::unique_ptr<RetryPolicy> policy = ....;
-   * Status status;
-   * while (!policy->IsExhausted()) {
-   *   auto response = try_rpc();  // typically `response` is StatusOr<T>
-   *   if (response.ok()) return response;
-   *   status = std::move(response).status();
-   *   if (!policy->OnFailure(response->status())) {
-   *     if (policy->IsPermanentFailure(response->status()) {
-   *       return StatusModifiedToSayPermanentFailureCausedTheProblem(status);
-   *     }
-   *     return StatusModifiedToSayPolicyExhaustionCausedTheProblem(status);
-   *   }
-   *   // sleep, which may exhaust the policy, even if it was not exhausted in
-   *   // the last call.
-   * }
-   * return StatusModifiedToSayPolicyExhaustionCausedTheProblem(status);
-   * @endcode
-   */
-  virtual bool OnFailure(Status const&) = 0;
-  virtual bool IsExhausted() const = 0;
-  virtual bool IsPermanentFailure(Status const&) const = 0;
-  //@}
-};
 
 /**
  * Trait based RetryPolicy.
