+++ google/cloud/retry_policy_test.cc
@@ -1,0 +1,40 @@
+// Copyright 2021 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "google/cloud/retry_policy.h"
+#include "google/cloud/status.h"
+#include <gmock/gmock.h>
+
+namespace google {
+namespace cloud {
+inline namespace GOOGLE_CLOUD_CPP_NS {
+namespace {
+
+struct Fake : RetryPolicy {
+  bool OnFailure(Status const&) override { return true; }
+  bool IsExhausted() const override { return true; }
+  bool IsPermanentFailure(Status const&) const override { return true; }
+};
+
+TEST(RetryPolicy, Fake) {
+  Fake fake;
+  EXPECT_TRUE(fake.OnFailure({}));
+  EXPECT_TRUE(fake.IsExhausted());
+  EXPECT_TRUE(fake.IsPermanentFailure({}));
+}
+
+}  // namespace
+}  // namespace GOOGLE_CLOUD_CPP_NS
+}  // namespace cloud
+}  // namespace google
